requirements.txt
asgiref==3.2.7
astroid==2.4.1
boto3==1.14.51
botocore==1.17.51
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-countries==6.1.2
django-crispy-forms==1.9.1
django-multiselectfield==0.1.12
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
gunicorn==20.0.4
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
psycopg2==2.8.5
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
